rename bridge to cgo
