+++ cli/cgo/bridge.go
@@ -1,4 +1,4 @@
-package bridge
+package cgo
 
 import (
 	"bytes"
