+++ cli/cgo/brightness.go
@@ -1,4 +1,4 @@
-package bridge
+package cgo
 
 import (
 	"log"
@@ -8,13 +8,13 @@ import (
 	"github.com/octu0/blurry"
 )
 
-func laplacianAction(c *cli.Context) error {
+func brightnessAction(c *cli.Context) error {
 	in, err := loadImage(c.String("input"))
 	if err != nil {
 		return err
 	}
 
-	out, err := blurry.Laplacian(in)
+	out, err := blurry.Brightness(in, c.Float64("factor"))
 	if err != nil {
 		return err
 	}
@@ -29,13 +29,18 @@ func laplacianAction(c *cli.Context) error {
 
 func init() {
 	addCommand(cli.Command{
-		Name:   "laplacian",
-		Action: laplacianAction,
+		Name:   "brightness",
+		Action: brightnessAction,
 		Flags: []cli.Flag{
 			cli.StringFlag{
 				Name:  "i,input",
 				Usage: "/path/to/input image",
 				Value: "./testdata/src.png",
+			},
+			cli.Float64Flag{
+				Name:  "f,factor",
+				Usage: "brightness factor",
+				Value: 1.5,
 			},
 		},
 	})
