+++ cli/cgo/canny.go
@@ -1,4 +1,4 @@
-package bridge
+package cgo
 
 import (
 	"log"
