+++ cli/cgo/highpass.go
@@ -1,4 +1,4 @@
-package bridge
+package cgo
 
 import (
 	"log"
@@ -8,13 +8,13 @@ import (
 	"github.com/octu0/blurry"
 )
 
-func contrastAction(c *cli.Context) error {
+func highpassAction(c *cli.Context) error {
 	in, err := loadImage(c.String("input"))
 	if err != nil {
 		return err
 	}
 
-	out, err := blurry.Contrast(in, c.Float64("factor"))
+	out, err := blurry.Highpass(in)
 	if err != nil {
 		return err
 	}
@@ -29,18 +29,13 @@ func contrastAction(c *cli.Context) error {
 
 func init() {
 	addCommand(cli.Command{
-		Name:   "contrast",
-		Action: contrastAction,
+		Name:   "highpass",
+		Action: highpassAction,
 		Flags: []cli.Flag{
 			cli.StringFlag{
 				Name:  "i,input",
 				Usage: "/path/to/input image",
 				Value: "./testdata/src.png",
-			},
-			cli.Float64Flag{
-				Name:  "f,factor",
-				Usage: "brightness factor",
-				Value: 0.525,
 			},
 		},
 	})
