+++ cli/cgo/invert.go
@@ -1,4 +1,4 @@
-package bridge
+package cgo
 
 import (
 	"log"
